split search and updates
