+++ src/containers/Search/Filters/Filters.module.scss
@@ -1,0 +1,40 @@
+@import '../../../components/containerGradient/saber/index.module.scss';
+
+.header {
+  // position: sticky;
+  // top: 110px;
+  // z-index: 10;
+  width: 70%;
+  display: flex;
+  margin: 10px auto -5px;
+  padding-left: 12px;
+  justify-content: space-between;
+  align-content: center;
+  // position: relative;
+
+  @include saber('blue', top);
+  padding-top: 17px;
+
+  padding-bottom: 20px;
+
+  // &::before {
+  //   border: 1px solid red;
+  //   content: '';
+  //   width: 100%;
+  //   left: 0;
+  //   height: 110%;
+
+  //   background-color: rgba(0, 0, 0, 0.3);
+  //   position: absolute;
+  //   z-index: -1;
+  //   // filter: blur(8px);
+  // }
+
+  .total {
+    color: rgba(221, 255, 255, 0.38);
+
+    span {
+      color: rgba(255, 255, 255, 0.78);
+    }
+  }
+}
