+++ src/containers/Search/constants.ts
@@ -1,0 +1,1 @@
+export const LIMIT = 15;
